Formalize project branding
There were several discussion ultimately culminating in the
set of guidelines included in this commit. The main motivation
here was the simplify the process of making new project logos.
Simplicity in this case comes at the cost of freedom, but
the overwhelming consensus was to favor simplicity and consistency
across all Nengo projects.

To make logo creation even easier, this commit includes a colour
selection script which outputs the colours that should be used
in different logo variants based on the primary colour chosen
by the project maintainer.

A subsequent commit will ensure that existing logos follow these
new guidelines.

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>